+++ src/src/lib/server/utils/users.ts
@@ -6,7 +6,7 @@ import type { ObjectId } from 'mongoose';
 
 import Users from '$db/schemas/User';
 import { generateUsername } from './auth';
-//import { sendWelcomeEmail } from './email';
+import { sendAdminWelcomeEmail } from './email';
 import { HASH_OPTIONS } from '$src/constants';
 import { upload, deleteFile } from '../storage';
 
@@ -59,7 +59,7 @@ export async function addUser(data: FormData, organisation: ObjectId | undefined
 	});
 
 	await newUser.save();
-	//sendWelcomeEmail(email, name, username);
+	sendAdminWelcomeEmail(email, name, username);
 
 	return { success: true };
 }
