+++ src/static/templates/lecturer-welcome.html
@@ -3,7 +3,7 @@
 	<head>
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Welcome to ClassConnect - Admin Account</title>
+		<title>Welcome to ClassConnect - Lecturer Account</title>
 		<style>
 			body {
 				font-family: Arial, sans-serif;
@@ -14,20 +14,20 @@
 				padding: 20px;
 			}
 			.header {
-				background-color: #27ae60;
+				background-color: #1e8449;
 				color: white;
 				text-align: center;
 				padding: 20px;
 				border-radius: 5px 5px 0 0;
 			}
 			.content {
-				background-color: #e8f8e8;
+				background-color: #e0f2e0;
 				padding: 20px;
 				border-radius: 0 0 5px 5px;
 			}
 			.button {
 				display: inline-block;
-				background-color: #2ecc71;
+				background-color: #27ae60;
 				color: white;
 				padding: 10px 20px;
 				text-decoration: none;
@@ -41,7 +41,7 @@
 				color: #666;
 			}
 			a {
-				color: #2ecc71;
+				color: #27ae60;
 				text-decoration: none;
 			}
 		</style>
@@ -54,8 +54,8 @@
 		<div class="content">
 			<h2>Congratulations, {{name}}! 🎉</h2>
 			<p>
-				Your application to work as an administrator for {{organization}} has been approved! We're
-				thrilled to have you join the ClassConnect community.
+				Your application to teach at {{organization}} has been approved! We're excited to have you
+				join our team of educators on the ClassConnect platform.
 			</p>
 			<h3>🔑 Your Account Details</h3>
 			<p>Your generated username is: <strong>{{username}}</strong></p>
@@ -68,21 +68,21 @@
 			<ol>
 				<li>Click the button below to sign in</li>
 				<li>Change your temporary password</li>
-				<li>Complete your administrator profile</li>
-				<li>Familiarize yourself with the ClassConnect platform</li>
+				<li>Complete your lecturer profile</li>
+				<li>Explore the ClassConnect teaching tools</li>
+				<li>Start creating your course content</li>
 			</ol>
 			<p>
-				As an administrator, you'll play a crucial role in managing the educational experience for
-				your organization. If you have any questions, don't hesitate to reach out to our support
-				team.
+				As a lecturer, you'll have access to powerful tools to create engaging course content and
+				interact with your students. If you need any assistance, our support team is here to help.
 			</p>
 			<a href="{{domain}}/signin" class="button">Sign In to Your Account</a>
-			<p>We're excited to see the positive impact you'll make in your new role! 🚀</p>
+			<p>We can't wait to see the knowledge you'll share and the students you'll inspire! 📚</p>
 			<p>Best regards,<br />The ClassConnect Team</p>
 		</div>
 		<div class="footer">
 			<p>
-				This email was sent to {{email}}. If you didn't apply for an administrator position at
+				This email was sent to {{email}}. If you didn't apply for a lecturer position at
 				ClassConnect, please contact our support team immediately.
 			</p>
 		</div>
