+++ src/static/templates/owner-welcome.html
@@ -3,7 +3,7 @@
 	<head>
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Welcome to ClassConnect - Admin Account</title>
+		<title>Welcome to ClassConnect - Owner Account</title>
 		<style>
 			body {
 				font-family: Arial, sans-serif;
@@ -14,20 +14,20 @@
 				padding: 20px;
 			}
 			.header {
-				background-color: #27ae60;
+				background-color: #2ecc71;
 				color: white;
 				text-align: center;
 				padding: 20px;
 				border-radius: 5px 5px 0 0;
 			}
 			.content {
-				background-color: #e8f8e8;
+				background-color: #f0fff0;
 				padding: 20px;
 				border-radius: 0 0 5px 5px;
 			}
 			.button {
 				display: inline-block;
-				background-color: #2ecc71;
+				background-color: #27ae60;
 				color: white;
 				padding: 10px 20px;
 				text-decoration: none;
@@ -41,7 +41,7 @@
 				color: #666;
 			}
 			a {
-				color: #2ecc71;
+				color: #27ae60;
 				text-decoration: none;
 			}
 		</style>
@@ -54,36 +54,33 @@
 		<div class="content">
 			<h2>Congratulations, {{name}}! 🎉</h2>
 			<p>
-				Your application to work as an administrator for {{organization}} has been approved! We're
-				thrilled to have you join the ClassConnect community.
+				Welcome to ClassConnect as the owner and special administrator! We're excited to have you on
+				board and help you create an amazing educational platform for your organization.
 			</p>
 			<h3>🔑 Your Account Details</h3>
 			<p>Your generated username is: <strong>{{username}}</strong></p>
-			<p>Your temporary password is: <strong>{{temp_password}}</strong></p>
 			<p>
-				Please use these credentials to log in to your account. You will be prompted to change your
-				password upon your first login.
+				You can use this username to log in to your account and start setting up your organization.
 			</p>
 			<h3>🌟 Next Steps</h3>
 			<ol>
-				<li>Click the button below to sign in</li>
-				<li>Change your temporary password</li>
-				<li>Complete your administrator profile</li>
-				<li>Familiarize yourself with the ClassConnect platform</li>
+				<li>Sign in to your account using the button below</li>
+				<li>Create your organization profile</li>
+				<li>Add other administrators, lecturers, and students to your organization</li>
+				<li>Customize your ClassConnect experience</li>
 			</ol>
 			<p>
-				As an administrator, you'll play a crucial role in managing the educational experience for
-				your organization. If you have any questions, don't hesitate to reach out to our support
-				team.
+				As the owner, you have full control over your organization's settings and user management.
+				If you need any assistance, our support team is always here to help.
 			</p>
 			<a href="{{domain}}/signin" class="button">Sign In to Your Account</a>
-			<p>We're excited to see the positive impact you'll make in your new role! 🚀</p>
+			<p>We're thrilled to see how you'll shape your educational community with ClassConnect! 🚀</p>
 			<p>Best regards,<br />The ClassConnect Team</p>
 		</div>
 		<div class="footer">
 			<p>
-				This email was sent to {{email}}. If you didn't apply for an administrator position at
-				ClassConnect, please contact our support team immediately.
+				This email was sent to {{email}}. If you didn't sign up for ClassConnect, please contact our
+				support team immediately.
 			</p>
 		</div>
 	</body>
