+++ src/static/templates/student-welcome.html
@@ -3,7 +3,7 @@
 	<head>
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Welcome to ClassConnect - Admin Account</title>
+		<title>Welcome to ClassConnect - Student Account</title>
 		<style>
 			body {
 				font-family: Arial, sans-serif;
@@ -14,14 +14,14 @@
 				padding: 20px;
 			}
 			.header {
-				background-color: #27ae60;
+				background-color: #16a085;
 				color: white;
 				text-align: center;
 				padding: 20px;
 				border-radius: 5px 5px 0 0;
 			}
 			.content {
-				background-color: #e8f8e8;
+				background-color: #d8f3e8;
 				padding: 20px;
 				border-radius: 0 0 5px 5px;
 			}
@@ -54,8 +54,8 @@
 		<div class="content">
 			<h2>Congratulations, {{name}}! 🎉</h2>
 			<p>
-				Your application to work as an administrator for {{organization}} has been approved! We're
-				thrilled to have you join the ClassConnect community.
+				Your application to attend {{organization}} has been approved! We're thrilled to welcome you
+				to our learning community on the ClassConnect platform.
 			</p>
 			<h3>🔑 Your Account Details</h3>
 			<p>Your generated username is: <strong>{{username}}</strong></p>
@@ -68,22 +68,23 @@
 			<ol>
 				<li>Click the button below to sign in</li>
 				<li>Change your temporary password</li>
-				<li>Complete your administrator profile</li>
-				<li>Familiarize yourself with the ClassConnect platform</li>
+				<li>Complete your student profile</li>
+				<li>Explore available courses</li>
+				<li>Connect with classmates and lecturers</li>
 			</ol>
 			<p>
-				As an administrator, you'll play a crucial role in managing the educational experience for
-				your organization. If you have any questions, don't hesitate to reach out to our support
-				team.
+				As a student, you'll have access to a wide range of courses, interactive learning tools, and
+				a supportive community. If you need any help getting started, our support team is always
+				here for you.
 			</p>
 			<a href="{{domain}}/signin" class="button">Sign In to Your Account</a>
-			<p>We're excited to see the positive impact you'll make in your new role! 🚀</p>
+			<p>We're excited to be part of your educational journey. Let's start learning! 📚</p>
 			<p>Best regards,<br />The ClassConnect Team</p>
 		</div>
 		<div class="footer">
 			<p>
-				This email was sent to {{email}}. If you didn't apply for an administrator position at
-				ClassConnect, please contact our support team immediately.
+				This email was sent to {{email}}. If you didn't apply to attend courses at ClassConnect,
+				please contact our support team immediately.
 			</p>
 		</div>
 	</body>
